looks good on ipad and all formats, added icons

+++ Source/Dot.h
@@ -20,7 +20,7 @@
 }
 
 @property int killNumber;
-@property int deathLevel;
+@property float deathLevel;
 @property Gameplay* gameplayLayer;
 
 -(void) randomizeValues;
